pyproject.toml
[tool.poetry]
name = "quill-sdk"
version = "0.2.3"
packages = [{include = "quill"}]
description = "A text-based adventure game engine with natural language processing"
authors = ["Timothy Pidashev <pidashev.tim@gmail.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/timmypidashev/quill"
keywords = ["game", "text-adventure", "interactive-fiction", "education"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "Topic :: Games/Entertainment :: Role-Playing",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
pyyaml = "^6.0"
colorama = "^0.4.4"
transformers = { version = "^4.20.0", optional = true }
torch = { version = "^1.10.0", optional = true }
openai = "^1.65.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^22.1.0"
flake8 = "^5.0.0"
mypy = "^0.950"
isort = "^5.10.1"

[tool.poetry.extras]
neural = ["transformers", "torch"]

[tool.poetry.scripts]
quill = "quill.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = "transformers.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "torch.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
